Components!!
- Fixed general components broken alignment and hitbox for ResponsiveLayout
- Added ImageComponent support for ResponsiveLayout

Everything working perfect :)

+++ src/main/java/codes/laivy/engine/utils/MathUtils.java
@@ -2,7 +2,7 @@ package codes.laivy.engine.utils;
 
 public class MathUtils {
 
-    public static float rthree(float a, float b, float c) {
+    public static double rthree(double a, double b, double c) {
         return (b * c) / a;
     }
